requirements.txt
# ============================================================================
# SWINGAI - PRODUCTION DEPENDENCIES
# Python 3.11+ Required
# ============================================================================

# ============================================================================
# WEB FRAMEWORK
# ============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3

# ============================================================================
# DATABASE & AUTHENTICATION
# ============================================================================
supabase==2.3.0
postgrest==0.13.2
realtime==1.0.4
storage3==0.7.4
gotrue>=2.4.0

# ============================================================================
# PAYMENTS
# ============================================================================
razorpay==1.4.1

# ============================================================================
# BROKER APIS
# ============================================================================
kiteconnect==4.2.0
requests==2.31.0
websocket-client==1.7.0

# ============================================================================
# MARKET DATA PROVIDERS
# ============================================================================
truedata>=7.0.0                # TrueData Velocity SDK (real-time + historical)
yfinance==0.2.35               # Fallback provider (15-20 min delayed)

# ============================================================================
# DATA PROCESSING
# ============================================================================
pandas==2.1.4
numpy==1.26.3
nsepy==0.8
ta==0.11.0
ta-lib==0.4.28
scipy==1.11.4

# ============================================================================
# PKSCREENER - Stock Screening
# https://github.com/pkjmesra/PKScreener
# ============================================================================
pkscreener

# ============================================================================
# MACHINE LEARNING
# ============================================================================
scikit-learn>=1.4.0
catboost>=1.2.2
torch>=2.1.2
pytorch-lightning>=2.1.3
pytorch-forecasting>=1.4.0
transformers>=4.36.2
google-genai==1.57.0

# ============================================================================
# TASK SCHEDULING
# ============================================================================
apscheduler==3.10.4
python-crontab==3.0.0

# ============================================================================
# CACHING & REAL-TIME
# ============================================================================
redis==5.0.1
python-redis-lock==4.0.0
aioredis==2.0.1

# ============================================================================
# SECURITY
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography>=41.0.7
pyjwt==2.8.0
pyotp>=2.9.0  # For TOTP in Angel One broker auth

# ============================================================================
# HTTP CLIENTS
# ============================================================================
httpx==0.24.1
aiohttp==3.9.1

# ============================================================================
# UTILITIES
# ============================================================================
python-dateutil==2.8.2
pytz==2023.3.post1
orjson==3.9.12

# ============================================================================
# MONITORING & LOGGING
# ============================================================================
sentry-sdk[fastapi]==1.39.2
python-json-logger==2.0.7

# ============================================================================
# TESTING
# ============================================================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0

# ============================================================================
# CODE QUALITY
# ============================================================================
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit==3.6.0
pylint==3.0.3

# ============================================================================
# DOCUMENTATION
# ============================================================================
mkdocs==1.5.3
mkdocs-material==9.5.3
